+++ main/src/main/java/net/oneandone/stool/cli/Remove.java
@@ -23,7 +23,7 @@ public class Remove extends StageCommand {
     private final boolean batch;
 
     public Remove(Session session, boolean batch) {
-        super(true, session, Mode.EXCLUSIVE, Mode.EXCLUSIVE);
+        super(session, Mode.EXCLUSIVE, Mode.EXCLUSIVE);
         this.batch = batch;
     }