+++ main/src/main/java/net/oneandone/stool/cli/Validate.java
@@ -50,7 +50,7 @@ public class Validate extends StageCommand {
     private Report report;
 
     public Validate(Session session, boolean email, boolean repair) {
-        super(false, session, Mode.SHARED, Mode.EXCLUSIVE);
+        super(session, Mode.SHARED, Mode.EXCLUSIVE);
         this.email = email;
         this.repair = repair;
     }