+++ CHANGELOG.md
@@ -48,7 +48,7 @@ Stages are now the central part; multiple project can be associated with the sam
   * source bind mounts (used for fitnesse) are gone
   * remove Project.updatesAvailable
   
-* use docker api 1.38
+* use docker api 1.39
 * dumped various 'auto' options - they were disabled for all remaining commands exception `remove`
 * start now takes an image index argument
 * added "images" command